hmm, planes again.. RMSD is weighted for weighted planes, so now both RMSDs are printed; rendering of 'regular' planes is fixed; the us of weights in the matching is to be implemented - for the esd calculation it is; also console text flickering when transparent objects shown/hidden is 'fixed' - the attributes are pushed before any rendering and popped upon completion...

+++ gxlib/gxapp.h
@@ -439,8 +439,8 @@ public:
 protected:  float ProbFactor(float Prob);
 public:     void CalcProbFactor(float Prob);
 
-  TXPlane *AddPlane(TXAtomPList& Atoms, bool Rectangular, int weightExtent=0);
-  TSPlane *TmpPlane(TXAtomPList* Atoms=NULL, int weightExtent=0); 
+  TXPlane *AddPlane(TXAtomPList& Atoms, bool Rectangular, double weightExtent=0);
+  TSPlane *TmpPlane(TXAtomPList* Atoms=NULL, double weightExtent=0); 
   void DeletePlane(TXPlane* plane);
   void ClearPlanes();
   TXPlane *XPlane(const olxstr& PlaneName);
